% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPower.R
\name{PlotPower}
\alias{PlotPower}
\title{Plot power based on estimate obtained from PowerIIphase function}
\usage{
PlotPower(x)
}
\arguments{
\item{x}{a PowerIIphase object from the PowerIIphase function.}
}
\value{
plot of powers based on different sampling designs and sample size of II phase
}
\description{
This function gives a plot power based on estimate obtained from objects of class PowerIIphase
}
\seealso{
\url{PowerIIphase} function for \code{x} term
}
